Use abstract class instead of sealed class
