+++ paymentsheet/src/main/java/com/stripe/android/paymentsheet/AnalyticEventInterceptor.kt
@@ -1,5 +1,7 @@
 package com.stripe.android.paymentsheet
 
+import com.stripe.android.paymentelement.AnalyticEventCallback
+
 internal object AnalyticEventInterceptor {
     var analyticEventCallback: AnalyticEventCallback? = null
 }
